+++ HISTORY.rst
@@ -8,6 +8,7 @@ LATEST_VERSION
 
 * Add command to import Events from GSX
 * Add command to import Organisations from GSX
+* Add command to import Person from GSX
 
 0.1.0 (2020-02-05)
 ---